requirements.txt
urllib3~=1.26.11
PyYAML~=6.0
selenium~=4.3.0